Support multiple ES-9s
Only the status component will show when multiple ES-9s are detected.
The user will be prompted to select which ES-9 input and output to use.

The names are expected to be the same for inputs and outputs which makes
it impossible to differentiate. Though the IDs seem to be unique and
durable for each ES-9. Users will have to know which input and output ID
correspond to the same ES-9.