Unrecognized MIDI messages are tolerated
Don't assume receiving unrecognized MIDI messages means the input device
is not an ES-9.

"Unrecogized" means there's no behavior tied to the message and
therefore the program has nothing to do. It was assumed that something
is wrong and the MIDI device should be cleared for the user to fix the
device selection. Any external MIDI message would therefore break the
application. This was realized by enabling "Mixer MIDI channels" and
sending control messages.

This change makes the application log the message and nothing else.